docs(components/*): add `<title>`, `<meta description/>` to component pages
