+++ app/docs/components/accordion/index.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import type { FC } from 'react';
+import AccordionDocs from './accordion.mdx';
+
+const AccordionPageContent: FC = () => <AccordionDocs />;
+
+export default AccordionPageContent;
