+++ app/docs/components/alert/index.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import type { FC } from 'react';
+import AlertDocs from './alert.mdx';
+
+const AlertPageContent: FC = () => <AlertDocs />;
+
+export default AlertPageContent;
