+++ app/docs/components/alert/page.tsx
@@ -1,17 +1,18 @@
-'use client';
-
-import type { FC } from 'react';
-
+import type { Metadata, NextPage } from 'next';
 import { DocsContentLayout } from '~/app/components/docs-content-layout';
-import AlertDocs from './alert.mdx';
+import AlertPageContent from '.';
 
-const AlertPage: FC = () => (
-  <DocsContentLayout
-    title="React Alert - Flowbite"
-    description="Show contextual information to your users using alert elements based on Tailwind CSS"
-  >
-    <AlertDocs></AlertDocs>
-  </DocsContentLayout>
-);
+export const metadata: Metadata = {
+  description: 'Show contextual information to your users using alert elements based on Tailwind CSS',
+  title: 'React Alert - Flowbite',
+};
+
+const AlertPage: NextPage = () => {
+  return (
+    <DocsContentLayout description={`${metadata.description}`} title={`${metadata.title}`}>
+      <AlertPageContent />
+    </DocsContentLayout>
+  );
+};
 
 export default AlertPage;
