+++ app/docs/components/avatar/index.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import type { FC } from 'react';
+import AvatarDocs from './avatar.mdx';
+
+const AvatarPageContent: FC = () => <AvatarDocs />;
+
+export default AvatarPageContent;
