+++ app/docs/components/badge/index.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import type { FC } from 'react';
+import BadgeDocs from './badge.mdx';
+
+const BadgePageContent: FC = () => <BadgeDocs />;
+
+export default BadgePageContent;
