+++ app/docs/components/breadcrumb/index.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import type { FC } from 'react';
+import BreadcrumbDocs from './breadcrumb.mdx';
+
+const BreadcrumbPageContent: FC = () => <BreadcrumbDocs />;
+
+export default BreadcrumbPageContent;
