+++ app/docs/components/breadcrumb/page.tsx
@@ -1,16 +1,18 @@
-'use client';
-
-import type { FC } from 'react';
+import type { Metadata, NextPage } from 'next';
 import { DocsContentLayout } from '~/app/components/docs-content-layout';
-import BreadcrumbDocs from './breadcrumb.mdx';
+import BreadcrumbPageContent from '.';
 
-const BreadcrumbPage: FC = () => (
-  <DocsContentLayout
-    title="React Breadcrumb - Flowbite"
-    description="Show the location of the current page in a hierarchical structure using the breadcrumb components"
-  >
-    <BreadcrumbDocs />
-  </DocsContentLayout>
-);
+export const metadata: Metadata = {
+  description: 'Show the location of the current page in a hierarchical structure using the breadcrumb components',
+  title: 'React Breadcrumb - Flowbite',
+};
+
+const BreadcrumbPage: NextPage = () => {
+  return (
+    <DocsContentLayout description={`${metadata.description}`} title={`${metadata.title}`}>
+      <BreadcrumbPageContent />
+    </DocsContentLayout>
+  );
+};
 
 export default BreadcrumbPage;
