+++ app/docs/components/button-group/index.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import type { FC } from 'react';
+import ButtonGroupDocs from './button-group.mdx';
+
+const ButtonGroupPageContent: FC = () => <ButtonGroupDocs />;
+
+export default ButtonGroupPageContent;
