+++ app/docs/components/button/index.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import type { FC } from 'react';
+import ButtonDocs from './button.mdx';
+
+const ButtonPageContent: FC = () => <ButtonDocs />;
+
+export default ButtonPageContent;
