+++ app/docs/components/card/index.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import type { FC } from 'react';
+import CardDocs from './card.mdx';
+
+const CardPageContent: FC = () => <CardDocs />;
+
+export default CardPageContent;
