+++ app/docs/components/carousel/index.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import type { FC } from 'react';
+import CarouselDocs from './carousel.mdx';
+
+const CarouselPageContent: FC = () => <CarouselDocs />;
+
+export default CarouselPageContent;
