+++ app/docs/components/carousel/page.tsx
@@ -1,16 +1,19 @@
-'use client';
-
-import type { FC } from 'react';
+import type { Metadata, NextPage } from 'next';
 import { DocsContentLayout } from '~/app/components/docs-content-layout';
-import CarouselDocs from './carousel.mdx';
+import CarouselPageContent from '.';
 
-const CarouselPage: FC = () => (
-  <DocsContentLayout
-    title="React Carousel - Flowbite"
-    description="Use the carousel component to slide through multiple elements and images using custom controls, indicators, intervals, and options"
-  >
-    <CarouselDocs />
-  </DocsContentLayout>
-);
+export const metadata: Metadata = {
+  description:
+    'Use the carousel component to slide through multiple elements and images using custom controls, indicators, intervals, and options',
+  title: 'React Carousel - Flowbite',
+};
+
+const CarouselPage: NextPage = () => {
+  return (
+    <DocsContentLayout description={`${metadata.description}`} title={`${metadata.title}`}>
+      <CarouselPageContent />
+    </DocsContentLayout>
+  );
+};
 
 export default CarouselPage;
