+++ app/docs/components/dropdown/index.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import type { FC } from 'react';
+import DropdownDocs from './dropdown.mdx';
+
+const DropdownPageContent: FC = () => <DropdownDocs />;
+
+export default DropdownPageContent;
