+++ app/docs/components/footer/index.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import type { FC } from 'react';
+import FooterDocs from './footer.mdx';
+
+const FooterPageContent: FC = () => <FooterDocs />;
+
+export default FooterPageContent;
