+++ app/docs/components/footer/page.tsx
@@ -1,16 +1,19 @@
-'use client';
-
-import type { FC } from 'react';
+import type { Metadata, NextPage } from 'next';
 import { DocsContentLayout } from '~/app/components/docs-content-layout';
-import FooterDocs from './footer.mdx';
+import CarouselPageContent from '.';
 
-const FooterPage: FC = () => (
-  <DocsContentLayout
-    title="React Footer - Flowbite"
-    description="Use the footer section at the bottom of every page to show valuable information to your users, such as sitemap links, a copyright notice, and a logo"
-  >
-    <FooterDocs />
-  </DocsContentLayout>
-);
+export const metadata: Metadata = {
+  description:
+    'Use the footer section at the bottom of every page to show valuable information to your users, such as sitemap links, a copyright notice, and a logo',
+  title: 'React Footer - Flowbite',
+};
+
+const FooterPage: NextPage = () => {
+  return (
+    <DocsContentLayout description={`${metadata.description}`} title={`${metadata.title}`}>
+      <CarouselPageContent />
+    </DocsContentLayout>
+  );
+};
 
 export default FooterPage;
