+++ app/docs/components/sidebar/index.tsx
@@ -1,0 +1,216 @@
+'use client';
+
+import type { FC } from 'react';
+import { BiBuoy } from 'react-icons/bi';
+import { HiArrowSmRight, HiChartPie, HiInbox, HiShoppingBag, HiTable, HiUser, HiViewBoards } from 'react-icons/hi';
+import { CodePreview } from '~/app/components/code-preview';
+import { DocsContentLayout } from '~/app/components/docs-content-layout';
+import { Badge, Sidebar } from '~/src';
+
+const SidebarPage: FC = () => (
+  <DocsContentLayout
+    title="React Sidebar - Flowbite"
+    description="Use the sidebar component to show a list of menu items and multi-level dropdown items on either side of the page to navigate on your website"
+  >
+    <p>
+      The sidebar component can be used as a complementary element relative to the navbar shown on either the left or
+      right side of the page used for the navigation on your web application, including menu items, multi-level dropdown
+      items, call to actions elements, and more.
+    </p>
+    <CodePreview title="Default sidebar" className="w-fit">
+      <Sidebar aria-label="Default sidebar example">
+        <Sidebar.Items>
+          <Sidebar.ItemGroup>
+            <Sidebar.Item href="#" icon={HiChartPie}>
+              Dashboard
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiViewBoards} label="Pro" labelColor="alternative">
+              Kanban
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiInbox} label="3">
+              Inbox
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiUser}>
+              Users
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiShoppingBag}>
+              Products
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiArrowSmRight}>
+              Sign In
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiTable}>
+              Sign Up
+            </Sidebar.Item>
+          </Sidebar.ItemGroup>
+        </Sidebar.Items>
+      </Sidebar>
+    </CodePreview>
+    <CodePreview title="Multi-level dropdown" className="w-fit">
+      <Sidebar aria-label="Sidebar with multi-level dropdown example">
+        <Sidebar.Items>
+          <Sidebar.ItemGroup>
+            <Sidebar.Item href="#" icon={HiChartPie}>
+              Dashboard
+            </Sidebar.Item>
+            <Sidebar.Collapse icon={HiShoppingBag} label="E-commerce">
+              <Sidebar.Item href="#">Products</Sidebar.Item>
+            </Sidebar.Collapse>
+            <Sidebar.Item href="#" icon={HiInbox}>
+              Inbox
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiUser}>
+              Users
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiShoppingBag}>
+              Products
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiArrowSmRight}>
+              Sign In
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiTable}>
+              Sign Up
+            </Sidebar.Item>
+          </Sidebar.ItemGroup>
+        </Sidebar.Items>
+      </Sidebar>
+    </CodePreview>
+    <CodePreview title="Content separator" className="w-fit">
+      <Sidebar aria-label="Sidebar with content separator example">
+        <Sidebar.Items>
+          <Sidebar.ItemGroup>
+            <Sidebar.Item href="#" icon={HiChartPie}>
+              Dashboard
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiViewBoards}>
+              Kanban
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiInbox}>
+              Inbox
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiUser}>
+              Users
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiShoppingBag}>
+              Products
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiArrowSmRight}>
+              Sign In
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiTable}>
+              Sign Up
+            </Sidebar.Item>
+          </Sidebar.ItemGroup>
+          <Sidebar.ItemGroup>
+            <Sidebar.Item href="#" icon={HiChartPie}>
+              Upgrade to Pro
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiViewBoards}>
+              Documentation
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={BiBuoy}>
+              Help
+            </Sidebar.Item>
+          </Sidebar.ItemGroup>
+        </Sidebar.Items>
+      </Sidebar>
+    </CodePreview>
+    <CodePreview title="CTA button" className="w-fit">
+      <Sidebar aria-label="Sidebar with call to action button example">
+        <Sidebar.Items>
+          <Sidebar.ItemGroup>
+            <Sidebar.Item href="#" icon={HiChartPie}>
+              Dashboard
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiViewBoards}>
+              Kanban
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiInbox}>
+              Inbox
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiUser}>
+              Users
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiShoppingBag}>
+              Products
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiArrowSmRight}>
+              Sign In
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiTable}>
+              Sign Up
+            </Sidebar.Item>
+          </Sidebar.ItemGroup>
+        </Sidebar.Items>
+        <Sidebar.CTA>
+          <div className="mb-3 flex items-center">
+            <Badge color="warning">Beta</Badge>
+            <button
+              aria-label="Close"
+              className="-m-1.5 ml-auto inline-flex h-6 w-6 rounded-lg bg-cyan-50 p-1 text-cyan-900 hover:bg-cyan-200 focus:ring-2 focus:ring-cyan-400 dark:bg-cyan-900 dark:text-cyan-400 dark:hover:bg-cyan-800"
+              type="button"
+            >
+              <svg
+                aria-hidden
+                className="h-4 w-4"
+                fill="currentColor"
+                viewBox="0 0 20 20"
+                xmlns="http://www.w3.org/2000/svg"
+              >
+                <path
+                  fillRule="evenodd"
+                  d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z"
+                  clipRule="evenodd"
+                />
+              </svg>
+            </button>
+          </div>
+          <p className="mb-3 text-sm text-cyan-900 dark:text-cyan-400">
+            Preview the new Flowbite dashboard navigation! You can turn the new navigation off for a limited time in
+            your profile.
+          </p>
+          <a
+            className="text-sm text-cyan-900 underline hover:text-cyan-800 dark:text-cyan-400 dark:hover:text-cyan-300"
+            href="#"
+          >
+            Turn new navigation off
+          </a>
+        </Sidebar.CTA>
+      </Sidebar>
+    </CodePreview>
+    <CodePreview title="Logo branding" className="w-fit">
+      <Sidebar aria-label="Sidebar with logo branding example">
+        <Sidebar.Logo href="#" img="/favicon.svg" imgAlt="Flowbite logo">
+          Flowbite
+        </Sidebar.Logo>
+        <Sidebar.Items>
+          <Sidebar.ItemGroup>
+            <Sidebar.Item href="#" icon={HiChartPie}>
+              Dashboard
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiViewBoards}>
+              Kanban
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiInbox}>
+              Inbox
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiUser}>
+              Users
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiShoppingBag}>
+              Products
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiArrowSmRight}>
+              Sign In
+            </Sidebar.Item>
+            <Sidebar.Item href="#" icon={HiTable}>
+              Sign Up
+            </Sidebar.Item>
+          </Sidebar.ItemGroup>
+        </Sidebar.Items>
+      </Sidebar>
+    </CodePreview>
+  </DocsContentLayout>
+);
+
+export default SidebarPage;
